Remove SSL keystore cli options (#15)
When we attempted to run with SSL and then disabled it, the cli options to
java to read a jks key store were left in bin/ycsb. This works fine against
non-ssl mongodb clusters, but when I connect to an Atlas cluster, this fails
because there is no file there.

Fix: remove the options as they are no longer used.